Add a few tests confirming the behavior of vCard library.
Fix test code as it doesn't convert multi-byte params to
UTF-8 properly.

Modified the content of v30_commpa_separated.vcf, as
the original intent of the file is not realistic
(I've never seen this kind of complicated form).
Also, without successfully parsing it, we torelate it.

Added comma-separated-types cases instead , as this might be
a bit more serious (a whole parse may fail due to
this form of types).

Change-Id: I1403e1c1d81d4325b25ad12ff143a76f0b9ab3a7