Stop using private APIs as much as possible.
- remove CharsetUtils stuffs
- remove private APIs in PhoneNumberUtils/TextUtils.

We keep the hidden flag "Data.FOR_EXPORT_ONLY", so this code still
needs access to private API. We don't remove it right now since we
may still need it while we confirmed that the other stuffs are "ok"
to be removed. Some more investigation will be needed in the future
to remove the flag.

Bug: 3143497
Change-Id: I074a7798d2208c3e14a524bd5f307a91ac50f735