Fix for malformed photo entry.
Blackberry vcard photo entry is not terminated with standard vcard 2.1
terminator. The vcard code is suppose to handle this irregularity already
but there was a bug where it would not work unless the next entry contained
a ':' right after the property name (i.e. "END:VCARD").  Cases with next
entries such as "TEL;TYPE=CELL:" was missed. Note the ":" is not right after
TEL.

Bug: 6979337
Change-Id: Ifa3dbb377b9da97ef7570c0f8eca8a1cb5e727a9