Can't send my wife's vcard between devices
Bug 13751984

If X-ANDROID-CUSTOM follows a photo's hex data, that line will get included
as part of the photo data and will cause Base64.decode() to throw an
exception. The code that gathers the hex data needs to finish if it sees
X-ANDROID-CUSTOM. Added a test case to test this "bad" vcard.

Change-Id: Ie60a1d09354a9c8a9dab603da01e9c7aff5eb070