Save away unknown properties that begin with X-.
PBAP sends over call logs using the vcard format and uses
the X-IRMC-DATETIME property to add in the timestamp of
the calls. The current VCardEntry drops entries that are
not recognized, and since these aren't recongized, they
are dropped.

Given that X- properties can be anything, this change saves
them away in a list that the caller can check and parse if
they desire to do so.

Change-Id: Iece345cc5b1f0ac3b7c8fbc29c78cc4deff11b22